+++ lib/src/analyzers/unused_code_analyzer/unused_code_config.dart
@@ -5,11 +5,13 @@ class UnusedCodeConfig {
   final Iterable<String> excludePatterns;
   final Iterable<String> analyzerExcludePatterns;
   final bool isMonorepo;
+  final bool shouldPrintConfig;
 
   const UnusedCodeConfig({
     required this.excludePatterns,
     required this.analyzerExcludePatterns,
     required this.isMonorepo,
+    required this.shouldPrintConfig,
   });
 
   /// Creates the config from analysis [options].
@@ -19,14 +21,17 @@ class UnusedCodeConfig {
         analyzerExcludePatterns:
             options.readIterableOfString(['analyzer', 'exclude']),
         isMonorepo: false,
+        shouldPrintConfig: false,
       );
 
   /// Creates the config from cli args.
   factory UnusedCodeConfig.fromArgs(
     Iterable<String> excludePatterns, {
     required bool isMonorepo,
+    required bool shouldPrintConfig,
   }) =>
       UnusedCodeConfig(
+        shouldPrintConfig: shouldPrintConfig,
         excludePatterns: excludePatterns,
         analyzerExcludePatterns: const [],
         isMonorepo: isMonorepo,
@@ -43,5 +48,6 @@ class UnusedCodeConfig {
           ...overrides.analyzerExcludePatterns,
         },
         isMonorepo: isMonorepo || overrides.isMonorepo,
+        shouldPrintConfig: shouldPrintConfig || overrides.shouldPrintConfig,
       );
 }