+++ lib/src/cli/commands/check_unused_code_command.dart
@@ -41,10 +41,12 @@ class CheckUnusedCodeCommand extends BaseCommand {
     final excludePath = argResults[FlagNames.exclude] as String;
     final reporterName = argResults[FlagNames.reporter] as String;
     final isMonorepo = argResults[FlagNames.isMonorepo] as bool;
+    final shouldPrintConfig = argResults[FlagNames.printConfig] as bool;
 
     final config = ConfigBuilder.getUnusedCodeConfigFromArgs(
       [excludePath],
       isMonorepo: isMonorepo,
+      shouldPrintConfig: shouldPrintConfig,
     );
 
     final unusedCodeResult = await _analyzer.runCliAnalysis(