+++ test/src/analyzers/unused_files_analyzer/unused_files_analyzer_test.dart
@@ -52,4 +52,5 @@ UnusedFilesConfig _createConfig({
       excludePatterns: const [],
       analyzerExcludePatterns: analyzerExcludePatterns,
       isMonorepo: false,
+      shouldPrintConfig: false,
     );