+++ test/src/analyzers/unused_l10n_analyzer/unused_l10n_analyzer_test.dart
@@ -154,4 +154,5 @@ UnusedL10nConfig _createConfig({
       excludePatterns: const [],
       analyzerExcludePatterns: analyzerExcludePatterns,
       classPattern: classPattern,
+      shouldPrintConfig: false,
     );